Add string interning to traced.
When a pointer is encountered for the first time, it is assigned an ID,
and a string registration message is sent to traced. The trace messages
then subsequently use the string ID rather than the string content,
reducing the SHM storage requirements (and also, coincidentally,
removing the limitation on string size).

This commit is not perfect: the registration ID is global, and currently
unlocked. We may want to use atomics for this instead.